changed inputs and design

+++ include/digitalInput.h
@@ -6,8 +6,9 @@ class DigitalInput : public Input
     using Input::Input;
 
 private:
-    inline uint16_t readInput(uint8_t pinnr)
+    inline uint8_t readInput(uint8_t pinnr)
     {
+
         return digitalRead(pinnr);
     }
 };
